requirements.txt
requests~=2.21.0
Pillow~=6.2.1
selenium~=3.141.0
twilio~=6.46.0
APScheduler~=3.6.3